requirements_torch.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astor=0.8.1=py37h06a4308_0
asttokens=2.0.5=pyhd8ed1ab_0
autograd=1.3=py_0
autograd-gamma=0.5.0=pyh9f0ad1d_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blosc=1.19.0=hd408876_0
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py37h06a4308_0
charls=2.1.0=he6710b0_2
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.5.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd3eb1b0_0
ecos=2.0.7.post1=py37h6323ea4_0
executing=0.6.0=pyhd8ed1ab_0
expat=2.3.0=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
formulaic=0.2.3=pyhd8ed1ab_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.9.0=pyhd3eb1b0_0
future=0.18.2=py37_1
giflib=5.1.4=h14c3975_1
glib=2.68.1=h36276a3_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py37hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
icecream=2.1.0=pyhd8ed1ab_0
icu=58.2=he6710b0_3
imagecodecs=2020.5.30=py37h567f118_1
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.2.0=h06a4308_610
interface_meta=1.2.3=pyhd8ed1ab_0
ipython=7.22.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libaec=1.0.4=he6710b0_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.0=h27cfd23_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libzopfli=1.0.3=he6710b0_0
lifelines=0.25.11=pyhd8ed1ab_0
llvmlite=0.36.0=py37h612dafd_4
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h2531618_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py37h27cfd23_1
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.1=py37ha9443f7_2
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
networkx=2.5=py_0
nibabel=3.2.1=pyhd8ed1ab_0
ninja=1.10.2=hff7bd54_1
numba=0.53.1=py37ha9443f7_0
numexpr=2.7.3=py37h22e1b3c_1
numpy=1.20.1=py37h93e21f0_0
numpy-base=1.20.1=py37h7d8b39e_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
osqp=0.5.0=py37h5f7f5f7_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py37h2531618_0
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
patsy=0.5.1=py37_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pydicom=2.1.2=pyhd3deb0d_0
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytorch=1.7.0=py3.7_cuda11.0.221_cudnn8.0.3_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.24.1=py37ha9443f7_0
scikit-survival=0.15.0.post0=py37hb93dfd8_0
scipy=1.6.2=py37had2a1c9_1
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
setuptools=52.0.0=py37h06a4308_0
shap=0.37.0=py37h10a2094_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
slicer=0.0.7=pyhd8ed1ab_0
snappy=1.1.8=he6710b0_0
sqlite=3.35.4=hdfb4753_0
statsmodels=0.12.2=py37h27cfd23_0
tabulate=0.8.9=py37h06a4308_0
tbb=2020.3=hfd86e86_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.3.31=pyhd3eb1b0_1
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.0=py37
torchvision=0.8.0=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py37h7b6447c_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
